Fix issue with mediaelement reload
This commit fixes an issue where if a user selected a new node in the
treeview and the content was refreshed, the mediaelement player would
not load and an image would load in it's place. To correct this, I have
had to stop using the auto-load class and attach the mediaelement js to
the js behaviours so it can be reloaded.